nvme: move security commands into a plugin
There are too many top-level commands cluttering the space. Move the
security commands into a plugin, e.g.,

  nvme security-send -> nvme security send
